workaround for xblock student data bug
The xblock architecture has a bug whereby students navigating within a
subsection see xblocks reinitialized with outdated (cached) student
data.  This workaround stores student data in the DOM and checks there
first on load to get the most current data.  This commit adds grading
to the information that gets stored in the DOM.